requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
torch
torchvision
transformers==4.48.0
accelerate
requests
python-dotenv
opencv-python-headless